pyproject.toml
[tool.poetry]
name = "lunchmoney-t212"
version = "0.1.0"
description = ""
authors = ["Alin Ali Hassan <alin.hassan@booking.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lunchable = "^1.3.2"
python-trading212-alinalihassan = "^2.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
